refactor: Move v1 Pet controller into its own file
